Testflight version 1.0.3 Removed duplicate label name in ui Removed DEPLOYMENT_TARGET from cmake Fixed data type cast for cmsOpenProfileFromMem Fixed data type for selectedItems()

+++ monitor.cpp
@@ -559,7 +559,7 @@ MonitorPrivate::statusChanged(Job::Status status)
 void
 MonitorPrivate::selectionChanged()
 {
-    int selectionCount = ui->items->selectedItems().count();
+    qint64 selectionCount = ui->items->selectedItems().count();
     if (selectionCount > 0) {
         if (selectionCount == 1) {
             QTreeWidgetItem* item = ui->items->selectedItems().first();
